refactor: moved client codecs and traits into selium-std
