+++ client/Cargo.toml
@@ -18,7 +18,6 @@ categories.workspace = true
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
-bincode = { version = "1.3", optional = true }
 bytes = "1.5"
 chrono = { version = "0.4", optional = true, default-features = false, features = [
     "clock",
@@ -29,48 +28,55 @@ rustls = "0.21"
 rustls-pemfile = "1.0"
 selium-common = { version = "0.1", path = "../common" }
 selium-std = { version = "0.1", optional = true, path = "../standard" }
-selium-traits = { version = "0.1", path = "../traits" }
-serde = { version = "1.0", optional = true, features = ["derive"] }
 tokio = { version = "1.32", features = ["full"] }
+
+[dev-dependencies]
+serde = "1.0"
 
 [features]
 chrono = ["dep:chrono"]
-bincode = ["dep:bincode", "dep:serde"]
-std = ["dep:selium-std"]
+std-compression = ["selium-std/compression"]
+std-traits = ["selium-std/traits"]
+std-codec = ["selium-std/codec"]
+std = ["std-compression", "std-codec"]
+default = ["std-traits"]
 
 [[example]]
 name = "publish"
 path = "examples/publish.rs"
+required-features = ["std-codec"]
 
 [[example]]
 name = "subscribe"
 path = "examples/subscribe.rs"
+required-features = ["std-codec"]
 
 [[example]]
 name = "publish_multitasking"
 path = "examples/publish_multitasking.rs"
+required-features = ["std-codec"]
 
 [[example]]
 name = "publish_bincode"
 path = "examples/publish_bincode.rs"
-required-features = ["bincode"]
+required-features = ["std-codec"]
 
 [[example]]
 name = "subscribe_bincode"
 path = "examples/subscribe_bincode.rs"
-required-features = ["bincode"]
+required-features = ["std-codec"]
 
 [[example]]
 name = "publish_chrono"
 path = "examples/publish_chrono.rs"
-required-features = ["chrono"]
+required-features = ["std-codec", "chrono"]
 
 [[example]]
 name = "compression"
 path = "examples/compression.rs"
-required-features = ["std"]
+required-features = ["std-codec", "std-compression"]
 
 [[example]]
 name = "decompression"
 path = "examples/decompression.rs"
-required-features = ["std"]
+required-features = ["std-codec", "std-compression"]
