+++ client/src/streams/subscriber.rs
@@ -1,4 +1,4 @@
-use crate::traits::{MessageDecoder, Open, Operations, Retain, SeliumCodec, TryIntoU64};
+use crate::traits::{Open, Operations, Retain, TryIntoU64};
 use crate::{StreamBuilder, StreamCommon};
 use anyhow::Result;
 use async_trait::async_trait;
@@ -7,7 +7,8 @@ use futures::{SinkExt, Stream, StreamExt};
 use quinn::Connection;
 use selium_common::protocol::{Frame, SubscriberPayload};
 use selium_common::types::BiStream;
-use selium_traits::compression::Decompress;
+use selium_std::traits::codec::MessageDecoder;
+use selium_std::traits::compression::Decompress;
 use std::marker::PhantomData;
 use std::pin::Pin;
 use std::sync::Arc;
@@ -67,10 +68,7 @@ impl<D, Item> Retain for StreamBuilder<SubscriberWantsOpen<D, Item>> {
     }
 }
 
-impl<D, Item> Operations for StreamBuilder<SubscriberWantsOpen<D, Item>>
-where
-    D: SeliumCodec,
-{
+impl<D, Item> Operations for StreamBuilder<SubscriberWantsOpen<D, Item>> {
     fn map(mut self, module_path: &str) -> Self {
         self.state.common.map(module_path);
         self
