+++ client/src/traits/mod.rs
@@ -1,9 +1,7 @@
 //! A collection of traits used by `Selium` and end-users.
 
-mod codec;
 mod stream;
 mod try_into_u64;
 
-pub use codec::*;
 pub use stream::*;
 pub use try_into_u64::*;
