+++ standard/src/codecs/bincode_codec.rs
@@ -1,4 +1,4 @@
-use crate::traits::{MessageDecoder, MessageEncoder, SeliumCodec};
+use crate::traits::codec::{MessageDecoder, MessageEncoder};
 use anyhow::Result;
 use bytes::{Buf, Bytes, BytesMut};
 use serde::{de::DeserializeOwned, Serialize};
@@ -50,8 +50,6 @@ impl<Item: DeserializeOwned> MessageDecoder<Item> for BincodeCodec<Item> {
         Ok(bincode::deserialize_from(buffer.reader())?)
     }
 }
-
-impl<Item> SeliumCodec for BincodeCodec<Item> {}
 
 #[cfg(test)]
 mod tests {
