+++ standard/src/compression/mod.rs
@@ -9,7 +9,7 @@ mod test {
     use bytes::Bytes;
     use fake::faker::lorem::en::Sentence;
     use fake::Fake;
-    use selium_traits::compression::{Compress, CompressionLevel, Decompress};
+    use crate::traits::compression::{Compress, CompressionLevel, Decompress};
 
     fn generate_payload() -> Bytes {
         let payload: String = Sentence(0..1).fake();
