+++ standard/src/compression/zstd/decomp.rs
@@ -1,6 +1,6 @@
+use crate::traits::compression::Decompress;
 use anyhow::Result;
 use bytes::Bytes;
-use selium_traits::compression::Decompress;
 
 #[derive(Debug)]
 pub struct ZstdDecomp;
