+++ standard/src/lib.rs
@@ -1,1 +1,6 @@
+#[cfg(feature = "codec")]
+pub mod codecs;
+#[cfg(feature = "compression")]
 pub mod compression;
+#[cfg(feature = "traits")]
+pub mod traits;
