+++ standard/src/traits/codec.rs
@@ -1,8 +1,6 @@
 use anyhow::Result;
 use bytes::{Bytes, BytesMut};
 
-pub(crate) trait SeliumCodec {}
-
 /// Provides an `encode` method for implementors to build their own encoder types.
 ///
 /// See [codecs](crate::codecs) for more information.
