+++ standard/src/traits/compression.rs
