+++ standard/src/traits/mod.rs
@@ -1,1 +1,2 @@
+pub mod codec;
 pub mod compression;
