+++ tests/Cargo.toml
@@ -6,5 +6,5 @@ edition = "2021"
 [dev-dependencies]
 anyhow = "1.0"
 futures = "0.3"
-selium = { path = "../client" }
+selium = { path = "../client", features = ["std"] }
 tokio = { version = "1.32", features = ["macros"] }
