--- traits/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "selium-traits"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anyhow = "1.0"
-bytes = "1.4"
